Retrieval-Augmented Generation (RAG) pipeline/RAG/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
requests>=2.31.0
pyyaml>=6.0.1
streamlit>=1.28.0
pytest>=7.4.3
numpy>=1.24.3
dotenv